// src/shared/api/generated/models/aIProposalsListParams.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { AIProposalsListStatusItem } from "./aIProposalsListStatusItem";
import type { AIProposalsListKindItem } from "./aIProposalsListKindItem";
import type { AIProposalsListPriorityItem } from "./aIProposalsListPriorityItem";
import type { AIProposalsListRead } from "./aIProposalsListRead";

export type AIProposalsListParams = {
  status?: AIProposalsListStatusItem[];
  kind?: AIProposalsListKindItem[];
  priority?: AIProposalsListPriorityItem[];
  read?: AIProposalsListRead;
  endUserId?: string;
  createdFrom?: string;
  createdTo?: string;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
  /**
   * @maxLength 2048
   */
  cursor?: string;
};